anchor enr rx to correct enr + adj sort on future coverage

+++ apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/coverage/converter/StringToDoubleConverter.java
@@ -18,7 +18,7 @@ public class StringToDoubleConverter implements AttributeConverter<Double, Strin
     }
 
     try {
-      return (double) Integer.parseInt(value.trim());
+      return Double.parseDouble(value.trim());
     } catch (NumberFormatException e) {
       throw new RuntimeException("Numeric input was not in a valid format: " + value, e);
     }
